attachment utils
