+++ src/components/Attachment/index.ts
@@ -3,3 +3,4 @@ export * from './AttachmentActions';
 export * from './Audio';
 export * from './Card';
 export * from './FileAttachment';
+export * from './utils';
